Add getting wallet balance hook
